19.07.11
위경도 insert 완료!!/지도 뿌려주고 지도 첨부 삭제
검색 기능 제대로수정할 것

+++ src/main/java/com/kitri/godinator/model/BbsDto.java
@@ -11,7 +11,7 @@ public class BbsDto extends BoardDto{
 	private String orginalName;		// 원본파일이름
 	private String savedName;		// 저장파일이름
 	private int fileSize;			// 파일크기
-	private int latitude;			// 위도
-	private int longitude;			// 경도
+	private float latitude;			// 위도
+	private float longitude;			// 경도
 
 }
